Create Q3
A climber is trying to reach a flag that is some height above the ground. In the attempt to reach the flag, the climber can make any number of jumps up the rock wall where the flag is mounted. Movements can only be made up the wall, and the climber must end exactly at the height of the flag.
 There are 2 types of jumps:
    1) A jump of height 1
    2) A jump of height bigJump
 Determine the minimum number of jumps it will take to reach the flag's exact height.

Example:
flagHeight = 8
bigJump = 3

The climber starts at height 0, takes two jumps of height bigJump and two of height 1 to reach exactly 8 units in 4 jumps 
Function Description: 
Complete the function jumps in the editor below.
jumps has the following parameter(s):
 int flagHeight: an int, the flag height 
 int bigJump: an int, the height of the second type of jump
Returns:
  int: an int, the minimum number of jumps necessary 
  
Input format:
 1st line contains int flagHeight
 2nd line contains int bigJump

Sample case:
3 -> flagHeight = 3
1 -> bigJump = 1